+++ mycodo/mycodo_flask/utils/utils_conditional.py
@@ -42,6 +42,10 @@ from mycodo.mycodo_client import DaemonControl
 control = DaemonControl()
 message = ''
 
+# The following functions are used to test the Conditional code.
+# The functions that are used in the production environment will
+# return the proper values and affect the proper systems when called.
+
 def measure(condition_id):
     # pylint: disable=unused-argument
     return random.choice([
@@ -66,9 +70,9 @@ def measure_dict(condition_id):
         {'time': '2019-04-24T18:12:00.000Z', 'value': 100000},
     ]
 
-def run_all_actions(message=message):
+def get_pid_setpoint(condition_id):
     # pylint: disable=unused-argument
-    pass
+    return True
 
 def run_action(action_id, message=message):
     # pylint: disable=unused-argument